+++ src/index.ts
@@ -3,6 +3,7 @@ import { parseCommand } from "./parser";
 import exitWithError from "./exit-with-error.js";
 import { exitWithErrorInternal } from "./exit-with-error-internal.js";
 import type { YaclilOptions, Command } from "../types.js";
+import { helpOption } from "./pre.js";
 
 /**
  * The YACLIL API
@@ -10,6 +11,9 @@ import type { YaclilOptions, Command } from "../types.js";
  */
 export function yaclil(rootCommand: Command, cliName: string, options?: YaclilOptions): void | never {
     options ??= {};
+    const globalOptions = options.globalOptions ?? [
+        helpOption,
+    ];
     const argv = options.customArgv ?? process.argv;
     // call the parser
     const parser = parseCommand({
@@ -17,16 +21,11 @@ export function yaclil(rootCommand: Command, cliName: string, options?: YaclilOp
         rootCommandName: cliName,
         initOptions: options,
         // remove the first two arguments
-        argv: argv.slice(2)
+        argv: argv.slice(2),
+        globalOptions,
     });
-    const helpCalled = parser.helpCalled;
     const parsed = parser.parsedObject;
-    // get the configured value of include help feature, or `true` by default
-    const helpConfigValue = parsed.command.helpOption ?? options.helpOptions ?? true;
-    if (helpCalled && helpConfigValue) {
-        // return the help string
-        console.log(renderHelp(parsed));
-    } else if (parser.standaloneOptionCalled) {
+    if (parser.standaloneOptionCalled) {
         // if standalone option called
         const parsedForStandalone = parser.parsedStandaloneOption;
         const handler = parsedForStandalone.parsedOption.option.standaloneHandler;