+++ src/pre.ts
@@ -1,3 +1,4 @@
 import treeGraphCommand from "./pre/tree-graph-command.js";
+import helpOption from "./pre/help-option.js";
 
-export { treeGraphCommand };
+export { treeGraphCommand, helpOption };